requirements.txt
python_speech_features
torchaudio==0.11.0 
webrtcvad
numpy==1.21.6

